refactor(metrics): split into modules
